+++ tests/testthat/test-s_parameter_to_xml.R
@@ -7,6 +7,16 @@ test_that("minimal use", {
   )
 })
 
+test_that("Numeric value 'Inf' becomes text 'Infinity'", {
+  s_param <- create_s_param(id = 1)
+  s_param$upper <- Inf
+  text <- s_parameter_to_xml(
+    s_param,
+    beauti_options = create_beauti_options()
+  )
+  expect_match(text, "Infinity")
+})
+
 test_that("reproduce same as in BEAUti v2.4 file", {
   xml <- readr::read_lines(beautier::get_beautier_path("tn93_2_4.xml"))
   expected <- unindent(stringr::str_subset(xml, "RealParameter.4.*name=.S."))